state at pipey s9

+++ reorg_files.py
@@ -3,6 +3,14 @@ import os
 import pandas as pd
 #%%
 
+import argparse
+parser = argparse.ArgumentParser()
+parser.add_argument("song")
+args = parser.parse_args()
+
+song = args.song
+
+
 excel_dir = r'G:\My Drive\AI-Art'
 
 fn = 'input_data.xlsx'
@@ -13,8 +21,8 @@ df = pd.read_excel(fp, sheet_name='transitions_pipey')
 df
 # %%
 
-movie_dir = r"G:\My Drive\AI-Art\pipey\transitions_2"
-output_basedir = r"G:\My Drive\AI-Art\pipey\scenes"
+movie_dir = r"G:\My Drive\AI-Art\{}\transitions".format(song)
+output_basedir = r"G:\My Drive\AI-Art\{}\scenes".format(song)
 
 #%%
 
@@ -46,8 +54,9 @@ for i, row in df.iterrows():
     output_fp = os.path.join(output_basedir, scene_dir, output_name)
 
     input_fp = os.path.join(movie_dir, output_name)
-
-    shutil.move(input_fp, output_fp)
+    
+    if os.path.exists(input_fp):
+        shutil.move(input_fp, output_fp)
